tls_client/requirements.txt
pymodbus[all]==3.6.8
python-decouple==3.8